\name{DataFreq}

\alias{DataFreq}

\docType{data}

\title{Conjunto de Dados de Frequencia}

\description{Conjunto de Dados de Frequencia, sobre consumo de cafes.}

\usage{data(DataFreq)}

\format{Conjunto de dados com 6 linhas e 9 colunas. Sendo 6 observacoes descritas por 9 variaveis: Grupo por sexo e idade,  Paris - Cafe Bourbon,	NY - Cafe Bourbon,	Tokyo - Cafe Bourbon,	NY - Cafe Acaia,	Tokyo - Cafe Acaia,	Paris - Cafe Catuai,	NY - Cafe Catuai,	 Tokyo - Cafe Catuai.}

\note{Este conjunto de dados foi extraido de (GREENACRE; BLASIUS, 2006) e os titulos das colunas foram adaptados a cafe.}

\references{GREENACRE, M.; BLASIUS, J.; \emph{Multiple Correspondence Analysis and Related Methods}. 1th. ed. New York: Taylor and Francis Group, 2006. 607 p.}

\examples{

data(DataFreq)
DataFreq
}

\keyword{Conjunto de Dados}
