\name{DataMix}

\alias{DataMix}

\docType{data}

\title{Conjunto de dados mistos}

\description{Conjunto simulado de dados mistos, sobre consumo de cafes.}

\usage{data(DataMix)}

\format{Conjunto de dados com 10 linhas e 7 colunas. Sendo 10 observacoes descritas por 7 variaveis: Cooperativas/Degustadores,  Medias das notas dadas aos cafes analisados,	Anos de trabalho como degustador,	Degustador com formacao tecnica,	Degustador com dedicacao exclusiva,	Frequencia media dos cafes classificados como especiais,	Frequencia media dos cafes classificados como comercias.}

\author{
Paulo Cesar Ossani

Marcelo Angelo Cirillo
}

\examples{
data(DataMix)
DataMix
}

\keyword{Conjunto de dados}
