\name{LocLab}

\alias{LocLab}

\title{Funcao para melhor posicao dos rotulos nos graficos.}

\description{
Funcao para melhor posicao dos rotulos nos graficos.
}

\usage{
LocLab(x, y = NULL, labels = seq(along = x), cex = 1,
       method = c("SANN", "GA"), allowSmallOverlap = FALSE,
       trace = FALSE, shadotext = FALSE, 
       doPlot = TRUE, \dots)
}

\arguments{
  \item{x}{Coordenada x}
  \item{y}{Coordenada y}
  \item{labels}{Os rotulos}
  \item{cex}{cex}
  \item{method}{Nao usado}
  \item{allowSmallOverlap}{Booleana}
  \item{trace}{Booleana}
  \item{shadotext}{Booleana}
  \item{doPlot}{Booleana}
  \item{\dots}{Outros argumentos passados para ou a partir de outros metodos}
}

\value{
Veja o texto da funcao.
}

