\name{MVar.pt}

\alias{MVar.pt}

\alias{MVar.pt}

\docType{package}

\title{Analise Multivariada (Brazilian Portuguese)}

\description{Pacote para Analise Multivariada, que possui funcoes que executam Analise de Correspondencia Simples (CA) e Multipla (MCA), Analise de Componentes Principais (PCA), Analise de Multiplos Fatores (MFA) para dados Quantitativos, Qualitativos, Frequencia (MFACT) e dados Mistos. Tambem possuim outras funcoes uteis para a Analise Multivariada.
}

\details{
\tabular{ll}{
Package: \tab MVar.pt\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-11-24\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}

\author{
Paulo Cesar Ossani

Marcelo Angelo Cirillo

Maintainer: Paulo Cesar Ossani <ossanipc@hotmail.com>
}

\references{
ABDESSEMED, L. and ESCOFIER, B.; Analyse factorielle multiple de tableaux de frequencies: comparaison avec l'analyse canonique des correspondences. \emph{Journal de la Societe de Statistique de Paris}, 137, p. 3-18, 1996.

ABDI, H; Singular Value Decomposition (SVD) and Generalized Singular Value Decomposition (GSVD). In:Salkind NJ, \emph{ed. Encyclopedia of Measurement and Statistics}. Thousand Oaks, CA: Sage; 2007.

ABDI, H; VALENTIN, D; Multiple factor analysis (MFA). In:Salkind NJ, \emph{ed. Encyclopedia of Measurement and Statistics}. Thousand Oaks, CA: Sage; 2007, 657-663.

ABDI, H; WILLIAMS, L; Principal Component Analysis.\emph{WIREs Comp Stat} 2010 2 433-459 DOI: 10.1002/wics.101

ABDI, H; WILLIAMS, L; VALENTIN, D; Multiple factor analysis: principal component analysis for multitable and multiblock data sets. \emph{WIREs Comput Stat} 2013. doi: 10.1002/wics.1246

BECUE-BERTAUT, M.; PAGES,J.; A principal axes method for comparing contingency tables: MFACT. \emph{Computational Statistics & Data Analysis}, v. 45, p. 481-503, 2004.

BECUE-BERTAUT, M.; PAGES,J.; Multiple factor analysis and clustering of a mixture of quantitative, categorical and frequency data. \emph{Computational Statistics & Data Analysis}, v. 52, p. 3255-3268, 2008.

ESCOFIER, B,; PAGES,J.; \emph{Analyse Factorielles Simples et Multiples}. Paris: Dunod., 1990.

ESCOFIER, B,; PAGES,J.; \emph{Analyses factorielles simples et multiples: objectifs, methodes et interpretation}. 4th. ed. Paris: Dunod, 2008. 318 p.

ESCOFIER, B,; PAGES,J.; Multiple Factor Analysis (AFUMULT package) \emph{Computational Statistics & Data Analysis}, v. 18, p. 121-140, 1994.

ESCOFIER, B. and DROUET, D.; Analyse des differences entre plusieurs tableaux de frequence. \emph{Les Cahiers de l'Analyse des Donnees},8 n. 4, p. 491-499, 1983.

ESCOFIER, B; Analyse factorielle en reference a un module: application a l'analyse d'un tableau d'echanges. \emph{Revue de Statistique Appliquee}, vol. 32, n. 4, p. 25-36, 1984.

GREENACRE, M.; BLASIUS, J.; \emph{Multiple Correspondence Analysis and Related Methods}. 1th. ed. New York: Taylor and Francis Group, 2006. 607 p.

HOTELLING, H.; Analysis of a complex of statistical variables into principal components. \emph{Journal of Educational Psychology}, v. 24, P. 417-441, set. 1933:

PAGES, J. Analyse factorielle multiple appliquee e aux variables qualitatives et aux donnees mixtes. \emph{Revue de Statistique Appliquee}. 50 (4), 5-37, 2002.

PAGES, J; Multiple Factor Analysis: main features and application to sensory data. \emph{Revista Colombiana de Estadistica}. vol. 27 n. 1, p. 1-26, 2004.

OSSANI, P. C.; CIRILLO, M. A.; Habilidades Sensoriais de Grupos Heterogeneos de Consumidores de Cafes Especiais Discriminadas pelo Metodo MFACT. in: XIII ENCONTRO MINEIRO DE ESTATISTICA (MGEST), 13., 2014, Diamantina. \emph{Anais}...Diamantina: UFVJM, 2014.

OSSANI, P. C. et al.; Multiplos fatores em analise de tabela de contingencia: Uma aplicacao na analise sensorial da qualidade de cafes especiais. in: 59 REUNIAO ANUAL DA REGIAO BRASILEIRA DA SOCIEDADE INTERNACIONAL DE BIOMETRIA (RBRAS), 59., 2014, Ouro Preto. \emph{Anais}...Ouro Preto: UFOP, 2014.

MINGOTI, SUELI APARECIDA; \emph{Analise de dados atraves de metodos de estatistica multivariada: uma abordagem aplicada}. Belo Horizonte: Editora UFMG, 2005. 297 p.

RENCHER, A.C.; \emph{Methods of Multivariate Analysis.} 2th. ed. New York: J.Wiley, 2002. 708 p.
}

\keyword{Analise Multivariada}
