\name{DataFreq}

\alias{DataFreq}

\docType{data}

\title{Conjunto de Dados de Frequencia}

\description{Conjunto simulado de dados com a frequencia semanal do numero de chicaras de cafes consumidas semanalmente em algumas capitais mundiais.}

\usage{data(DataFreq)}

\format{Conjunto de dados com 6 linhas e 9 colunas. Sendo 6 observacoes descritas por 9 variaveis: Grupo por sexo e idade, Sao Paulo - Cafe Bourbon, Londres - Cafe Bourbon, Atenas - Cafe Bourbon, Londres - Cafe Acaia, Atenas - Cafe Acaia, Sao Paulo - Cafe Catuai, Londres - Cafe Catuai, Atenas - Cafe Catuai.

}

\author{
Paulo Cesar Ossani

Marcelo Angelo Cirillo
}

\examples{
data(DataFreq)
DataFreq
}

\keyword{Conjunto de Dados}
