\name{DataQuali}

\alias{DataQuali}

\docType{data}

\title{Conjunto simulado de dados qualitativos.}

\description{Conjunto de dados qualitativos, sobre consumo de cafes.}

\usage{data(DataQuali)}

\format{Conjunto de dados com 12 linhas e 6 colunas. Sendo 12 observacoes descritas  por 6 variaveis: Sexo,  Idade,	Fumante,	Estado Civil,	Esportista,	Estuda.}

\author{
Paulo Cesar Ossani

Marcelo Angelo Cirillo
}

\examples{
data(DataQuali)
DataQuali
}

\keyword{Conjunto de dados}
