\name{DataQuan}

\alias{DataQuan}

\docType{data}

\title{Conjunto de Dados Quantitativos}

\description{Conjunto de Dados Quantitativos, sobre consumo de cafes.}

\usage{data(DataQuan)}

\format{Conjunto de dados com 6 linhas e 11 colunas. Sendo 6 observacoes descritas por 11 variaveis: Cafes, Achocolatado, Acaramelado,	Maduro,	Doce,	Delicado,	Amendoado, Acaramelado,	Achocolatado, Picante, Acaramelado.}

\note{Este conjunto de dados foi extraido de (ABDI; WILLIAMS, 2010) e os titulos das colunas foram adaptados a cafe.}

\references{ABDI, H.; WILLIAMS, L. Principal component analysis. \emph{WIREs Computational Statatistics}, New York, v. 2, n. 4, p. 433-459, July/Aug. 2010.}

\examples{
data(DataQuan)
DataQuan
}

\keyword{Conjunto de Dados}
