\name{DataQuan}

\alias{DataQuan}

\docType{data}

\title{Conjunto de Dados Quantitativos}

\description{Conjunto simulado de dados Quantitativos, sobre notas dadas a algumas caracteristas sensoriais dos cafes.}

\usage{data(DataQuan)}

\format{Conjunto de dados com 6 linhas e 11 colunas. Sendo 6 observacoes descritas por 11 variaveis: Cafes, Achocolatado, Acaramelado,	Maduro,	Doce,	Delicado,	Amendoado, Acaramelado,	Achocolatado, Picante, Acaramelado.}

\author{
Paulo Cesar Ossani

Marcelo Angelo Cirillo
}

\examples{
data(DataQuan)
DataQuan
}

\keyword{Conjunto de Dados}
